requirements.txt
mpi4py>=3
numpy>=1.23
Pillow>=9
PyYAML>=6.0
scikit_learn>=1.3
spacy>=3.6
torch>=2
torchvision>=0.15
tqdm>=4.65.0
